requirements.txt
python-dotenv>=1.0.0
matplotlib>=3.7.1